frankenstein-tool/requirements.txt
moviepy>=2.1.1
Pillow>=10.4.0
numpy>=2.1.0
pytest>=8.3.0
